# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Monitoring TMAT - Portal V1",
    "description": "Mock backend API for plantation monitoring system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/portal_v1/site": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Site",
        "description": "Get sites with optional filtering by location or company.\n\nParameters:\n- id_kecamatan: Filter sites by kecamatan/district ID\n- id_kelurahan: Filter sites by kelurahan/village ID\n- id_perusahaan: Filter sites by company ID",
        "operationId": "get_site_api_portal_v1_site_get",
        "parameters": [
          {
            "name": "id_kecamatan",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kecamatan/district ID",
              "title": "Id Kecamatan"
            },
            "description": "Filter by kecamatan/district ID"
          },
          {
            "name": "id_kelurahan",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kelurahan/village ID",
              "title": "Id Kelurahan"
            },
            "description": "Filter by kelurahan/village ID"
          },
          {
            "name": "id_perusahaan",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by company ID",
              "title": "Id Perusahaan"
            },
            "description": "Filter by company ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SiteResponse"
                  },
                  "title": "Response Get Site Api Portal V1 Site Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal_v1/site/kecamatan": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Kecamatan",
        "description": "Get all kecamatan (districts) or filter by kabupaten.\n\nParameters:\n- id_kabupaten: Filter kecamatan by parent kabupaten/regency ID",
        "operationId": "get_kecamatan_api_portal_v1_site_kecamatan_get",
        "parameters": [
          {
            "name": "id_kabupaten",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kabupaten/regency ID",
              "title": "Id Kabupaten"
            },
            "description": "Filter by kabupaten/regency ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KecamatanResponse"
                  },
                  "title": "Response Get Kecamatan Api Portal V1 Site Kecamatan Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal_v1/site/kelurahan": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Kelurahan",
        "description": "Get all kelurahan (villages) or filter by kecamatan.\n\nParameters:\n- id_kecamatan: Filter kelurahan by parent kecamatan/district ID",
        "operationId": "get_kelurahan_api_portal_v1_site_kelurahan_get",
        "parameters": [
          {
            "name": "id_kecamatan",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kecamatan/district ID",
              "title": "Id Kecamatan"
            },
            "description": "Filter by kecamatan/district ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KelurahanResponse"
                  },
                  "title": "Response Get Kelurahan Api Portal V1 Site Kelurahan Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal_v1/site/kabupaten": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Kabupaten",
        "description": "Get all kabupaten (regencies) or filter by provinsi.\n\nParameters:\n- id_provinsi: Filter kabupaten by parent provinsi/province ID",
        "operationId": "get_kabupaten_api_portal_v1_site_kabupaten_get",
        "parameters": [
          {
            "name": "id_provinsi",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by provinsi/province ID",
              "title": "Id Provinsi"
            },
            "description": "Filter by provinsi/province ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KabupatenResponse"
                  },
                  "title": "Response Get Kabupaten Api Portal V1 Site Kabupaten Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal_v1/site/provinsi": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Provinsi",
        "description": "Get all provinces (provinsi)",
        "operationId": "get_provinsi_api_portal_v1_site_provinsi_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProvinsiResponse"
                  },
                  "title": "Response Get Provinsi Api Portal V1 Site Provinsi Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KabupatenResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_kabupaten": {
            "type": "integer",
            "title": "Id Kabupaten"
          },
          "id_provinsi": {
            "type": "integer",
            "title": "Id Provinsi"
          },
          "nama_kabupaten": {
            "type": "string",
            "title": "Nama Kabupaten"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_kabupaten",
          "id_provinsi",
          "nama_kabupaten"
        ],
        "title": "KabupatenResponse",
        "description": "Schema for regency/kabupaten response"
      },
      "KecamatanResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_kecamatan": {
            "type": "integer",
            "title": "Id Kecamatan"
          },
          "id_kabupaten": {
            "type": "integer",
            "title": "Id Kabupaten"
          },
          "nama_kecamatan": {
            "type": "string",
            "title": "Nama Kecamatan"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_kecamatan",
          "id_kabupaten",
          "nama_kecamatan"
        ],
        "title": "KecamatanResponse",
        "description": "Schema for district/kecamatan response"
      },
      "KelurahanResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_kelurahan": {
            "type": "integer",
            "title": "Id Kelurahan"
          },
          "id_kecamatan": {
            "type": "integer",
            "title": "Id Kecamatan"
          },
          "nama_kelurahan": {
            "type": "string",
            "title": "Nama Kelurahan"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_kelurahan",
          "id_kecamatan",
          "nama_kelurahan"
        ],
        "title": "KelurahanResponse",
        "description": "Schema for village/kelurahan response"
      },
...
      "ProvinsiResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_provinsi": {
            "type": "integer",
            "title": "Id Provinsi"
          },
          "nama_provinsi": {
            "type": "string",
            "title": "Nama Provinsi"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_provinsi",
          "nama_provinsi"
        ],
        "title": "ProvinsiResponse",
        "description": "Schema for province/provinsi response"
      },
...
      "SiteResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_perusahaan": {
            "type": "integer",
            "title": "Id Perusahaan"
          },
          "nama_site": {
            "type": "string",
            "title": "Nama Site"
          },
          "id_provinsi": {
            "type": "integer",
            "title": "Id Provinsi"
          },
          "id_kabupaten": {
            "type": "integer",
            "title": "Id Kabupaten"
          },
          "id_kecamatan": {
            "type": "integer",
            "title": "Id Kecamatan"
          },
          "id_kelurahan": {
            "type": "integer",
            "title": "Id Kelurahan"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "keterangan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keterangan"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_perusahaan",
          "nama_site",
          "id_provinsi",
          "id_kabupaten",
          "id_kecamatan",
          "id_kelurahan",
          "latitude",
          "longitude",
          "created_at"
        ],
        "title": "SiteResponse",
        "description": "Schema for master_site response"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}